[SDWebImage] Update to version 3.5.4

+++ SDWebImage/samples/SDWebImageMTDialogSample/SDWebImageMTDialogSample/DetailViewController.cs
@@ -38,7 +38,7 @@ namespace SDWebImageMTDialogSample
 			}
 		}
 
-		void ProgressHandler (uint receivedSize, long expectedSize)
+		void ProgressHandler (int receivedSize, int expectedSize)
 		{
 			if (activityIndicator == null) {
 				InvokeOnMainThread (()=> {
